Remove validation for username/password
The default value for `authentication` in the configuration model is `form`.

This is not actually used unless it is explicitly defined at this stage, but because there is validation ensuring they are defined if the value for authentication mode is set to an "enabled" value, instantiating the default configuration model fails.

For now, remove this validation. It might be re-added in the future if a way to enforce unmanaged default values is added to Buildarr in the future.